Moved clipboard/pastebuffer functionality to new file clipboard.py
Also:
- Converted global can_clip variable to an instance attribute of cmd2.Cmd class
